PLANET-4811 Create single file for all submenu functions
