+++ assets/src/blocks/Submenu/submenuFunctions.js
@@ -1,3 +1,18 @@
+// Map for old attribute 'submenu_style'
+const SUBMENU_STYLES = {
+  1: 'long',
+  2: 'short',
+  3: 'sidebar'
+};
+
+export const getSubmenuStyle = (className, submenu_style) => {
+  let style = submenu_style ? SUBMENU_STYLES[submenu_style] : 'long';
+  if (className && className.includes('is-style-')) {
+    style = className.split('is-style-')[1];
+  }
+  return style;
+};
+
 export const addSubmenuActions = submenu => {
   if (submenu && Array.isArray(submenu)) {
     for (let i = 0; i < submenu.length; i++) {
